+++ includes/functions/updates/index.php
@@ -1,0 +1,1 @@
+<?php // quiet.
